requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==20.1.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23.1
tzdata==2022.5
urllib3==1.26.12